+++ packages/ts-sdk/src/types/types.ts
@@ -1040,7 +1040,7 @@ export type WriteSetChangeDeleteTableItem = {
   state_key_hash: string;
   handle: string;
   key: string;
-  data?: DeletedTableData;
+  data?: DeletedTableData | null;
 };
 
 /**
@@ -1072,7 +1072,7 @@ export type WriteSetChangeWriteTableItem = {
   handle: string;
   key: string;
   value: string;
-  data?: DecodedTableData;
+  data?: DecodedTableData | null;
 };
 
 /**